% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llcont.R
\name{llcont.coxph}
\alias{llcont.coxph}
\title{Getting partial log-likelihood of a \code{coxph} object for individual cases}
\usage{
\method{llcont}{coxph}(x)
}
\arguments{
\item{x}{\code{coxph} model with \code{x=T}}
}
\value{
a vector of individual likelihoods (sorted by decreasing event times)
}
\description{
Getting partial log-likelihood of a \code{coxph} object for individual cases
}
\details{
ties are handled according to Efron
}
\examples{
\dontrun{
### example data set from Fine paper, section 5
require("survival")
mod <- coxph(Surv(time, status==2) ~ age + log(albumin) + log(bili) + edema + log(protime), data=subset(pbc, !is.na(trt)),  x=T)
# individual LLs sum up to total model LL
sum(llcont(mod));logLik(mod)[1]
}
}
\author{
Thomas Hielscher
}
