% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llcont.R
\name{llcont}
\alias{llcont}
\title{Getting partial log-likelihood of a \code{coxph} object for individual cases}
\usage{
llcont(x, ...)
}
\arguments{
\item{model}{\code{coxph} model with \code{x=T}}
}
\value{
a vector of individual likelihoods, individuals are sorted by event time
}
\description{
Getting partial log-likelihood of a \code{coxph} object for individual cases
}
\details{
ties are handled according to Efron
}
\examples{
\dontrun{
### example data set from Fine paper, section 5
require("randomForestSRC")
data(pbc)
pbc <- subset(pbc, !is.na(treatment))
mod <- coxph(Surv(days, status) ~ age + albumin + bili + edema + prothrombin, data=pbc,  x=T)
# individual LLs sum up to total model LL
sum(llcont(mod));logLik(mod)[1]
}
#' @importFrom sandwich estfun
}
\author{
Thomas Hielscher
}
