% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plrtest.R
\name{plrtest}
\alias{plrtest}
\title{Partial Likelihood Ratio Test for non-nested \code{coxph} models}
\usage{
plrtest(object1, object2, nested = FALSE, adjusted = FALSE)
}
\arguments{
\item{object1}{\code{coxph} model with \code{x=T}}

\item{object2}{\code{coxph} model with \code{x=T}, based on the identical data and order of observations as for \code{object1}}

\item{nested}{specify if models are nested, default ist \code{FALSE}}

\item{adjusted}{specify if test statistic for non-nested models should be adjusted for different complexities of the models using BIC-type adjustment (p/2)log n - (q/2)log n, default ist \code{FALSE}}
}
\value{
a object of class \code{finetest}
}
\description{
Partial Likelihood Ratio Test for non-nested \code{coxph} models
}
\examples{
\dontrun{
### example data set from Fine paper, section 5
require("survival")
pbc  <- subset(pbc, !is.na(trt))
mod1 <- coxph(Surv(time, status==2) ~ age, data=pbc, x=T)
mod2 <- coxph(Surv(time, status==2) ~ age + albumin + bili + edema + protime, data=pbc,  x=T)
mod3 <- coxph(Surv(time, status==2) ~ age + log(albumin) + log(bili) + edema + log(protime), data=pbc, x=T)
plrtest(mod3, mod2, nested=F) # non-nested models
plrtest(mod3, mod1, nested=T) # nested models
}

}
\references{
Fine J.P., Comparing nonnested Cox models, Biometrika (2002), 89, 3, 635-647.

Vuong Q.H., Likelihood Ratio Tests for Model Selection and Non-Nested Hypotheses, Econometrika (1989), 57, 2, 307-333.

Merkle E.C. and You D., Testing Nonnested Structural Equation Models, Psychological Methods (2016), 21, 2, 151-163.

Edgar Merkle and Dongjun You (2018). nonnest2: Tests of Non-Nested Models. R package version 0.5-1.

P. Duchesne, P. Lafaye de Micheaux, Computing the distribution of quadratic forms: Further comparisons between the Liu-Tang-Zhang approximation and exact methods, Computational Statistics and Data Analysis, Volume 54, (2010), 858-862

Schwarz G. (1978). Estimating the Dimension of a Model. Annals of Statistics, 6, 461.464.
}
\author{
Thomas Hielscher
}
