% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plrtest.R
\name{plrtest}
\alias{plrtest}
\title{Partial Likelihood Ratio Test for non-nested \code{coxph} models}
\usage{
plrtest(object1, object2, nested = FALSE)
}
\arguments{
\item{object1}{\code{coxph} model with \code{x=T}}

\item{object2}{\code{coxph} model with \code{x=T}, based on the identical data and order of observations as for \code{object1}}

\item{nested}{specify if models are nested, default ist \code{FALSE}}
}
\value{
a object of class \code{finetest}
}
\description{
Partial Likelihood Ratio Test for non-nested \code{coxph} models
}
\examples{
\dontrun{
### example data set from Fine paper, section 5
require("randomForestSRC")
data(pbc)
pbc  <- subset(pbc, !is.na(treatment)) # 312 randomized patients
mod1 <- coxph(Surv(days, status) ~ age, data=pbc, ties = "efron", x=T)
mod2 <- coxph(Surv(days, status) ~ age + log(albumin) + log(bili) + edema + log(prothrombin), data=pbc, x=T)
mod3 <- coxph(Surv(days, status) ~ age + albumin + bili + edema + prothrombin, data=pbc,  x=T)
plrtest(mod1, mod3, nested=T) 
plrtest(mod2, mod3, nested=F) 
}
}
\references{
Fine J.P., Comparing nonnested Cox models, Biometrika (2002), 89, 3, 635-647.

Vuong Q.H., Likelihood Ratio Tests for Model Selection and Non-Nested Hypotheses, Econometrika (1989), 57, 2, 307-333.

Merkle E.C. and You D., Testing Nonnested Structural Equation Models, Psychological Methods (2016), 21, 2, 151-163.

P. Duchesne, P. Lafaye de Micheaux, Computing the distribution of quadratic forms: Further comparisons between the Liu-Tang-Zhang approximation and exact methods, Computational Statistics and Data Analysis, Volume 54, (2010), 858-862
}
\author{
Thomas Hielscher
}
